move duplicated functions from customLayout to layoutEngine
