+++ bigbluebutton-html5/imports/ui/components/layout/layout-manager/layoutEngine.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { layoutSelect, layoutSelectInput } from '/imports/ui/components/layout/context';
 import DEFAULT_VALUES from '/imports/ui/components/layout/defaultValues';
-import { LAYOUT_TYPE } from '/imports/ui/components/layout/enums';
+import { LAYOUT_TYPE, DEVICE_TYPE } from '/imports/ui/components/layout/enums';
 
 import CustomLayout from '/imports/ui/components/layout/layout-manager/customLayout';
 import SmartLayout from '/imports/ui/components/layout/layout-manager/smartLayout';
@@ -18,13 +18,20 @@ const LayoutEngine = ({ layoutType }) => {
   const notificationsBarInput = layoutSelectInput((i) => i.notificationsBar);
   const cameraDockInput = layoutSelectInput((i) => i.cameraDock);
   const presentationInput = layoutSelectInput((i) => i.presentation);
+  const actionbarInput = layoutSelectInput((i) => i.actionBar);
+  const sidebarNavigationInput = layoutSelectInput((i) => i.sidebarNavigation);
+  const sidebarContentInput = layoutSelectInput((i) => i.sidebarContent);
 
   const fullscreen = layoutSelect((i) => i.fullscreen);
   const isRTL = layoutSelect((i) => i.isRTL);
   const fontSize = layoutSelect((i) => i.fontSize);
+  const deviceType = layoutSelect((i) => i.deviceType);
 
+  const isMobile = deviceType === DEVICE_TYPE.MOBILE;
   const windowWidth = () => window.document.documentElement.clientWidth;
   const windowHeight = () => window.document.documentElement.clientHeight;
+  const min = (value1, value2) => (value1 <= value2 ? value1 : value2);
+  const max = (value1, value2) => (value1 >= value2 ? value1 : value2);
 
   const bannerAreaHeight = () => {
     const { hasNotification } = notificationsBarInput;
@@ -85,7 +92,7 @@ const LayoutEngine = ({ layoutType }) => {
       left: !isRTL ? mediaAreaBounds.left : 0,
       zIndex: 1,
     };
-  }
+  };
 
   const calculatesActionbarHeight = () => {
     const { actionBarHeight, actionBarPadding } = DEFAULT_VALUES;
@@ -100,11 +107,177 @@ const LayoutEngine = ({ layoutType }) => {
     };
   };
 
+  const calculatesActionbarBounds = (mediaAreaBounds) => {
+    const actionBarHeight = calculatesActionbarHeight();
+
+    return {
+      display: actionbarInput.hasActionBar,
+      width: mediaAreaBounds.width,
+      height: actionBarHeight.height,
+      innerHeight: actionBarHeight.innerHeight,
+      padding: actionBarHeight.padding,
+      top: windowHeight() - actionBarHeight.height,
+      left: !isRTL ? mediaAreaBounds.left : 0,
+      zIndex: 1,
+    };
+  };
+
+  const calculatesSidebarNavWidth = () => {
+    const {
+      sidebarNavMinWidth,
+      sidebarNavMaxWidth,
+    } = DEFAULT_VALUES;
+
+    const { isOpen, width: sidebarNavWidth } = sidebarNavigationInput;
+
+    let minWidth = 0;
+    let width = 0;
+    let maxWidth = 0;
+    if (isOpen) {
+      if (isMobile) {
+        minWidth = windowWidth();
+        width = windowWidth();
+        maxWidth = windowWidth();
+      } else {
+        if (sidebarNavWidth === 0) {
+          width = min(max((windowWidth() * 0.2), sidebarNavMinWidth), sidebarNavMaxWidth);
+        } else {
+          width = min(max(sidebarNavWidth, sidebarNavMinWidth), sidebarNavMaxWidth);
+        }
+        minWidth = sidebarNavMinWidth;
+        maxWidth = sidebarNavMaxWidth;
+      }
+    }
+    return {
+      minWidth,
+      width,
+      maxWidth,
+    };
+  };
+
+  const calculatesSidebarNavHeight = () => {
+    const { navBarHeight } = DEFAULT_VALUES;
+    const { isOpen } = sidebarNavigationInput;
+
+    let sidebarNavHeight = 0;
+    if (isOpen) {
+      if (isMobile) {
+        sidebarNavHeight = windowHeight() - navBarHeight - bannerAreaHeight();
+      } else {
+        sidebarNavHeight = windowHeight() - bannerAreaHeight();
+      }
+    }
+    return sidebarNavHeight;
+  };
+
+  const calculatesSidebarNavBounds = () => {
+    const { sidebarNavTop, navBarHeight, sidebarNavLeft } = DEFAULT_VALUES;
+
+    let top = sidebarNavTop + bannerAreaHeight();
+
+    if (isMobile) {
+      top = navBarHeight + bannerAreaHeight();
+    }
+
+    return {
+      top,
+      left: !isRTL ? sidebarNavLeft : null,
+      right: isRTL ? sidebarNavLeft : null,
+      zIndex: isMobile ? 11 : 2,
+    };
+  };
+
+  const calculatesSidebarContentWidth = () => {
+    const {
+      sidebarContentMinWidth,
+      sidebarContentMaxWidth,
+    } = DEFAULT_VALUES;
+
+    const { isOpen, width: sidebarContentWidth } = sidebarContentInput;
+
+    let minWidth = 0;
+    let width = 0;
+    let maxWidth = 0;
+
+    if (isOpen) {
+      if (isMobile) {
+        minWidth = windowWidth();
+        width = windowWidth();
+        maxWidth = windowWidth();
+      } else {
+        if (sidebarContentWidth === 0) {
+          width = min(
+            max((windowWidth() * 0.2), sidebarContentMinWidth), sidebarContentMaxWidth,
+          );
+        } else {
+          width = min(max(sidebarContentWidth, sidebarContentMinWidth),
+            sidebarContentMaxWidth);
+        }
+        minWidth = sidebarContentMinWidth;
+        maxWidth = sidebarContentMaxWidth;
+      }
+    }
+    return {
+      minWidth,
+      width,
+      maxWidth,
+    };
+  };
+
+  const calculatesSidebarContentBounds = (sidebarNavWidth) => {
+    const { navBarHeight, sidebarNavTop } = DEFAULT_VALUES;
+
+    let top = sidebarNavTop + bannerAreaHeight();
+
+    if (isMobile) top = navBarHeight + bannerAreaHeight();
+
+    let left = isMobile ? 0 : sidebarNavWidth;
+    let right = isMobile ? 0 : sidebarNavWidth;
+    left = !isRTL ? left : null;
+    right = isRTL ? right : null;
+
+    const zIndex = isMobile ? 11 : 1;
+
+    return {
+      top,
+      left,
+      right,
+      zIndex,
+    };
+  };
+
+  const calculatesMediaAreaBounds = (sidebarNavWidth, sidebarContentWidth) => {
+    const { navBarHeight } = DEFAULT_VALUES;
+    const { height: actionBarHeight } = calculatesActionbarHeight();
+    let left = 0;
+    let width = 0;
+    if (isMobile) {
+      width = windowWidth();
+    } else {
+      left = !isRTL ? sidebarNavWidth + sidebarContentWidth : 0;
+      width = windowWidth() - sidebarNavWidth - sidebarContentWidth;
+    }
+
+    return {
+      width,
+      height: windowHeight() - (navBarHeight + actionBarHeight + bannerAreaHeight()),
+      top: navBarHeight + bannerAreaHeight(),
+      left,
+    };
+  };
+
   const common = {
     bannerAreaHeight,
     baseCameraDockBounds,
     calculatesNavbarBounds,
     calculatesActionbarHeight,
+    calculatesActionbarBounds,
+    calculatesSidebarNavWidth,
+    calculatesSidebarNavHeight,
+    calculatesSidebarNavBounds,
+    calculatesSidebarContentWidth,
+    calculatesSidebarContentBounds,
+    calculatesMediaAreaBounds,
   };
 
   switch (layoutType) {
